TASK: Move the DimensionSwitcher to the PrimaryToolbar (#2587)
