+++ packages/neos-ui/src/Containers/PrimaryToolbar/DimensionSwitcher/index.spec.js
