+++ packages/neos-ui/src/Containers/PrimaryToolbar/style.css
@@ -1,20 +1,19 @@
 .primaryToolbar {
-    &:before,
-    &:after {
-        content: '';
-        display: table;
-    }
-    &:after {
-        clear: both;
-    }
+    display: flex;
 
     z-index: var(--zIndex-PrimaryToolbar);
 }
 .primaryToolbar--isHidden {
     display: none;
 }
+.primaryToolbar__leftSidedActions {
+    display: flex;
+    flex: 0 0 var(--size-SidebarWidth);
+}
+.primaryToolbar__actions {
+    display: flex;
+    flex: 1;
+}
 .primaryToolbar__rightSidedActions {
-    position: absolute;
-    right: 0;
-    top: 0;
+    display: flex;
 }
