--- packages/neos-ui/src/Containers/SecondaryToolbar/DimensionSwitcher/style.css
@@ -1,29 +1,0 @@
-.dropDown {
-    max-width: 352px;
-    border-right: 1px solid var(--colors-ContrastDark);
-}
-.dropDown__btnIcon {
-    margin-right: .5em;
-}
-
-.cancelButton {
-    margin-right: var(--spacing-Quarter) !important;
-}
-
-.dimensionLabel {
-    margin-top: var(--spacing-Half);
-    margin-bottom: var(--spacing-Quarter);
-}
-
-.buttonGroup {
-    text-align: right;
-    margin-top: var(--spacing-Half);
-}
-
-.selectPreset {
-    margin-right: var(--spacing-Full);
-}
-
-.dimmed {
-    filter: opacity(50%);
-}
